plugins: Eliminate unnecessary `run_npm.js`
I had anticipated more shared logic than we actually need (the
abstraction in `run_npm.js` is YAGNI).
